Reorganize twinkle client
