+++ crates/adapters/celestia/src/lib.rs
@@ -18,6 +18,6 @@ pub mod verifier;
 #[cfg(feature = "native")]
 pub use da_service::{CelestiaConfig, CelestiaService};
 #[cfg(feature = "native")]
-pub use twinkle::{Network, TwinkleBlockHeader, TwinkleClient, TwinkleConfig};
+pub use twinkle::{Network, TwinkleClient, TwinkleConfig};
 
 pub use crate::celestia::*;
